+++ forge20/src/main/java/com/vecoo/extraquests/command/ExtraQuestsCommand.java
@@ -45,7 +45,7 @@ public class ExtraQuestsCommand {
         UUID targetUUID = UtilPlayer.findUUID(target);
 
         if (targetUUID == null) {
-            source.sendSystemMessage(UtilChat.formatMessage(ExtraQuests.instance().localeConfig().playerNotFound()
+            source.sendSystemMessage(UtilChat.formatMessage(ExtraQuests.getInstance().getLocaleConfig().getPlayerNotFound()
                     .replace("%player%", target)));
             return 0;
         }
@@ -58,7 +58,7 @@ public class ExtraQuestsCommand {
             task.progress(teamData, key, amount, ignore);
         }
 
-        source.sendSystemMessage(UtilChat.formatMessage(ExtraQuests.instance().localeConfig().addKeyValue()
+        source.sendSystemMessage(UtilChat.formatMessage(ExtraQuests.getInstance().getLocaleConfig().getAddKeyValue()
                 .replace("%player%", target)
                 .replace("%key%", key)
                 .replace("%value%", String.valueOf(amount))));
@@ -66,10 +66,10 @@ public class ExtraQuestsCommand {
     }
 
     private static int executeReload(@NotNull CommandSourceStack source) {
-        ExtraQuests.instance().loadConfig();
-        ExtraQuests.instance().loadStorage();
+        ExtraQuests.getInstance().loadConfig();
+        ExtraQuests.getInstance().loadStorage();
 
-        source.sendSystemMessage(UtilChat.formatMessage(ExtraQuests.instance().localeConfig().reload()));
+        source.sendSystemMessage(UtilChat.formatMessage(ExtraQuests.getInstance().getLocaleConfig().getReload()));
         return 1;
     }
 }