+++ forge20/src/main/java/com/vecoo/extraquests/mixin/CommandRewardMixin.java
@@ -72,10 +72,10 @@ public abstract class CommandRewardMixin {
     )
     public int claim(Commands instance, CommandSourceStack source, String command) {
         if (this.console) {
-            ServerConfig config = ExtraQuests.instance().config();
+            ServerConfig config = ExtraQuests.getInstance().getConfig();
 
             if (config.isBlacklistConsole()) {
-                for (String blacklistCommand : config.blacklistConsoleList()) {
+                for (String blacklistCommand : config.getBlacklistConsoleList()) {
                     if (command.contains(blacklistCommand)) {
                         return 0;
                     }