+++ changelog.md
@@ -1,1 +1,3 @@
-* Update name methods 1.20.1 forge.
+* Change name methods.
+* Update 1.20.1 timer storage.
+* Update depends.